Split request mocks out across different files.
